Doubles To Floats
but you don't understand, i NEED the full 64 bits of precision!

+++ Source/GAME/Components/CPlayer.cs
@@ -7,11 +7,11 @@ namespace GAME.Components
 {
 	public class CPlayer : Component
 	{
-		public double maxSpeed = 0.75;
-		public double acceleration = 12.0;
-		public double friction = 1.0 - 0.05;
+		public float maxSpeed = 0.75f;
+		public float acceleration = 12.0f;
+		public float friction = 1.0f - 0.05f;
 
-		public double jumpVel = 1.5;
+		public float jumpVel = 1.5f;
 
 		CRigidbody rb;
 		Texture body;
@@ -27,7 +27,7 @@ namespace GAME.Components
 
 		public override void Update()
 		{
-			var moveInput = ((Input.GetButton(Inputs.D) ? 1.0 : 0.0) - (Input.GetButton(Inputs.A) ? 1.0 : 0.0));
+			var moveInput = ((Input.GetButton(Inputs.D) ? 1.0f : 0.0f) - (Input.GetButton(Inputs.A) ? 1.0f : 0.0f));
 
 			rb.velocity.x = rb.velocity.x * friction;
