+++ Source/MGE/Essentials/Perlin.cs
@@ -33,7 +33,7 @@ namespace MGE
 			Calculate();
 		}
 
-		public double Noise(double x, double y, double z)
+		public float Noise(float x, float y, float z)
 		{
 			int
 				X = (int)Math.Floor(x) & 255,
@@ -42,7 +42,7 @@ namespace MGE
 			x -= Math.Floor(x);
 			y -= Math.Floor(y);
 			z -= Math.Floor(z);
-			double u = Fade(x),
+			float u = Fade(x),
 			v = Fade(y),
 			w = Fade(z);
 			int
@@ -60,15 +60,15 @@ namespace MGE
 				Lerp(u, Gradient(p[AB + 1], x, y - 1, z - 1), Gradient(p[BB + 1], x - 1, y - 1, z - 1))));
 		}
 
-		double Lerp(double t, double a, double b) => a + t * (b - a);
+		float Lerp(float t, float a, float b) => a + t * (b - a);
 
-		double Fade(double t) =>
-		 t * t * t * (t * (t * 6d - 15d) + 10d);
+		float Fade(float t) =>
+		 t * t * t * (t * (t * 6f - 15f) + 10f);
 
-		double Gradient(int hash, double x, double y, double z)
+		float Gradient(int hash, float x, float y, float z)
 		{
 			int h = hash & 15;
-			double u = h < 8 ? x : y,
+			float u = h < 8 ? x : y,
 			v = h < 4 ? y : h == 12 || h == 14 ? x : z;
 			return ((h & 1) == 0 ? u : -u) + ((h & 2) == 0 ? v : -v);
 		}
